pyproject.toml
[project]
name = "pyscript-daydayup"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.8"
dependencies = [
    "config>=0.5.1",
    "pillow>=10.4.0",
    "pystray>=0.19.5",
    "win11toast>=0.35",
]
